fix tinytimer & revert to c++17
c++20 gh actions bug: github.com/actions/runner-images/issues/8659

+++ headeronly/tinytimer.hpp
@@ -14,7 +14,7 @@ public:
 
     ~Timer() {
         const auto m_end = std::chrono::high_resolution_clock::now();
-        const auto duration = std::chrono::duration_cast<std::chrono::microseconds>(m_end - m_start).count();
+        const auto duration = static_cast<long long>(std::chrono::duration_cast<std::chrono::microseconds>(m_end - m_start).count());
         std::printf("%s: ", m_name);
 
         if (duration > us_in_s) {
